+++ vault/CHANGELOG.md
@@ -1,3 +1,11 @@
+2.9
+
+* Add Nomad proxy
+* Allow Nomad PKI to issue server certs for nomad.service.consul
+* Allow Consul and Vault servers to request Nomad client certs
+
+---
+
 2.8
 
 * Version bump for new base image